Update the create, update and delete curriculum buttons to use the template components and a better handling of vuex logic

+++ music_paradigm_web/src/components/admin/curriculums/widget-manage-curriculum/section-editor/buttons/button-delete-curriculum.component.vue
@@ -1,25 +1,46 @@
 <template>
-    <button v-on:click="handleCurriculumDeletion" class="widget-button red" :class="isButtonActive || 'inactive'">
-        Delete Curriculum
-    </button>
+    <TemplateButtonComponent v-on:click="handleButtonPress" color="red" :isActive="isButtonActive"
+        :isLoading="isButtonLoading" :isFrozen="isButtonFrozen" v-bind="$attrs" text="Delete Curriculum" />
 </template>
 
 <script>
-import '@/styles/widget-template.css';
 import { mapActions, mapGetters } from 'vuex';
 
+import TemplateButtonComponent from '@/components/admin/template/template-button.component.vue';
+
 export default {
+    components: {
+        TemplateButtonComponent,
+    },
     computed: {
-        ...mapGetters('managementCurriculums/selection', ['curriculumSelectionId', 'hasSelectedCurriculum']),
+        ...mapGetters('managementCurriculums', [
+            'isDeletingCurriculum',
+            'isExecutingCurriculumCommand',
+        ]),
+        ...mapGetters('managementCurriculums/selection', [
+            'curriculumSelectionId',
+            'hasSelectedCurriculum',
+        ]),
+        isButtonFrozen() {
+            return !this.isDeletingCurriculum && this.isExecutingCurriculumCommand; // Has other user command running
+        },
         isButtonActive() {
             return this.hasSelectedCurriculum;
+        },
+        isButtonLoading() {
+            return this.isDeletingCurriculum;
         },
     },
     methods: {
         ...mapActions('managementCurriculums', ['deleteCurriculum']),
-        handleCurriculumDeletion() {
+        handleButtonPress() {
             if (!this.isButtonActive) return;
-            const answer = window.confirm('Are your sure you want to delete the curriculum?');
+            let answer = window.confirm('Are you sure you want to delete the curriculum? \nThis cannot be undone.');
+            if (!answer) return;
+            window.confirm(
+                'Have you made sure that no user was currently completing this curriculum?\n' +
+                'Are you sure you want to go ahead with the deletion of this curriculum?'
+            );
             if (answer) this.deleteCurriculum(this.curriculumSelectionId);
         },
     }
